README touchups
Including a demo on the spellaphone.org page.  And a 404 redirect
page. And so many favicons.

Switch over to using jsdeliver instead of rawgit.

And have the demo translate words to numbers and have some text about
it as well.

Get rid of fancy ES6 stuff so it'll work in more browsers.

And skip node 10 for now.