Add jsdelivr field
The "main" field points at index.js for Node usage. The "browser"
field tells the packager (Browserify) to replace data.js references
with min-data.js (optimized for browser use).

The actual combined file for use in a browser is spellaphone.js. This
sets that to be the default file on jsDelivr such that links like

  https://cdn.jsdelivr.net/npm/spellaphone@1.0.4

can be used.